+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Implement `serde::Serialize` and `serde::Deserialize` for `Duration` using `FromStr` and `Display` implementations, when feature `derive-serde` is enabled.
+
 ## concordium-contracts-common 6.0.0 (2023-05-08)
 
 - Remove the `Copy` requirement for deserialization of BTreeMap and BTreeSet.